refactor(runner): move timing.ts to utils directory
